Make scoreboard calculation functions pure
