+++ core/server-core/src/services/challenge/index.ts
@@ -21,7 +21,7 @@ import {
 } from "../../errors.ts";
 import type { SomeJSONSchema } from "ajv/dist/types/json-schema.js";
 import { SubmissionDAO } from "../../dao/submission.ts";
-import { CACHE_SCORE_NAMESPACE } from "../scoreboard.ts";
+import { CACHE_SCORE_NAMESPACE } from "../scoreboard/index.ts";
 import { ChallengePlugin } from "./types.ts";
 import { CoreChallengePlugin } from "./core_plugin.ts";
