+++ core/server-core/src/util/object.ts
@@ -19,3 +19,18 @@ export const partition = <T>(arr: T[], x: (t: T) => boolean) => {
   arr.forEach((a) => (x(a) ? truthy.push(a) : falsey.push(a)));
   return [truthy, falsey];
 };
+
+export const deepEqual = <T>(a: T, b: T): boolean => {
+  if (a === b) {
+    return true;
+  }
+
+  const bothAreObjects =
+    a && b && typeof a === "object" && typeof b === "object";
+
+  return Boolean(
+    bothAreObjects &&
+      Object.keys(a).length === Object.keys(b).length &&
+      Object.entries(a).every(([k, v]) => deepEqual(v, b[k as keyof T])),
+  );
+};
